# docs/swagger.yaml
basePath: /
definitions:
  domain.DailyExecutionCount:
    properties:
      count:
        type: integer
      date:
        type: string
    type: object
...
  domain.UserStatistics:
    properties:
      last_7_days:
        items:
          $ref: '#/definitions/domain.DailyExecutionCount'
        type: array
      today:
        type: integer
      total:
        type: integer
      user_id:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API для учета выполненных упражнений и получения статистики по пользователю.
  title: Workout Tracker API
  version: "1.0"
paths:
...
  /users/{userID}/statistics:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserStatistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить статистику пользователя
      tags:
      - statistics
swagger: "2.0"